Refactor plant gas mutation effects and localization
