+++ Content.Shared/EntityEffects/Effects/PlantMutateGases.cs
@@ -1,3 +1,4 @@
+using Robust.Shared.Localization;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Random;
 using System.Linq;
@@ -7,7 +8,7 @@ namespace Content.Shared.EntityEffects.Effects;
 /// <summary>
 ///     changes the gases that a plant or produce create.
 /// </summary>
-public sealed partial class PlantMutateExudeGasses : EventEntityEffect<PlantMutateExudeGasses>
+public sealed partial class PlantMutateExudeGasses : EntityEffect
 {
     [DataField]
     public float MinValue = 0.01f;
@@ -15,16 +16,23 @@ public sealed partial class PlantMutateExudeGasses : EventEntityEffect<PlantMuta
     [DataField]
     public float MaxValue = 0.5f;
 
+    public override void Effect(EntityEffectBaseArgs args)
+    {
+        // This is handled in EntityEffectSystem.cs
+    }
+
     protected override string? ReagentEffectGuidebookText(IPrototypeManager prototype, IEntitySystemManager entSys)
     {
-        return "TODO";
+        return Loc.GetString("reagent-effect-guidebook-plant-mutate-exude-gasses",
+            ("minValue", MinValue),
+            ("maxValue", MaxValue));
     }
 }
 
 /// <summary>
 ///     changes the gases that a plant or produce consumes.
 /// </summary>
-public sealed partial class PlantMutateConsumeGasses : EventEntityEffect<PlantMutateConsumeGasses>
+public sealed partial class PlantMutateConsumeGasses : EntityEffect
 {
     [DataField]
     public float MinValue = 0.01f;
@@ -32,8 +40,15 @@ public sealed partial class PlantMutateConsumeGasses : EventEntityEffect<PlantMu
     [DataField]
     public float MaxValue = 0.5f;
 
+    public override void Effect(EntityEffectBaseArgs args)
+    {
+        // This is handled in EntityEffectSystem.cs
+    }
+
     protected override string? ReagentEffectGuidebookText(IPrototypeManager prototype, IEntitySystemManager entSys)
     {
-        return "TODO";
+        return Loc.GetString("reagent-effect-guidebook-plant-mutate-consume-gasses",
+            ("minValue", MinValue),
+            ("maxValue", MaxValue));
     }
 }
